refactor(drive): useDriveFileSharePermissionStrings hook
